Database.Kafkar.Types

Topics

data Topic

Constructors

Topic

Fields

topicName :: !String

partition :: !Int

segments :: !(Vector Segment)

Segments sorted by initial offset

data Segment

Kafka streams are split into segments on disk. Each segment comprises a log file, which stores the data, and an index, which allows fast seeking to a specific message within the log file. The files associated with a segment are named according to the offset (within the whole topic) of the first message in the segment. The message offsets used in the index are given relative to the starting offset of the segment.

Constructors

Segment

Fields

initialOffset :: !Offset

index :: !Index

logPath :: !FilePath

Message sets

data MessageEntry

Constructors

MessageEntry

Fields

offset :: !Offset

size :: !Int32

message :: !Message

data Message

Constructors

Message

Fields

attributes :: !Attributes

key :: !(Maybe ByteString)

value :: !(Maybe ByteString)

data Attributes

Constructors

Attributes

Fields

compression :: !Codec

Indices

type Index = Vector IndexEntry

An unboxed vector of offset-position relations, which is

sorted in both columns; and

sparse in both columns.

data IndexEntry

Relates a logical message offset (given relative to the start of the segment) to a byte offset within the log file for the segment.

Constructors

IndexEntry

Fields

relativeOffset :: !RelativeOffset

logPosition :: !LogPosition

The byte position in the corresponding log file

Primitives

data Codec

Message compression algorithm

Constructors

None

GZip

Snappy

newtype Offset

The logical offset of a message within a topic

Constructors

Offset Int64

newtype RelativeOffset

The difference between two logical message offsets

Constructors

RelativeOffset Int32

newtype LogPosition

The byte-offset of a message within a log file

Constructors

LogPosition Int32